MiniProject 7

+++ angular_material/src/app/menu-management/menu-management.component.ts
@@ -19,6 +19,8 @@ export class MenuManagementComponent implements OnInit {
   private subs = new SubSink();
   menu: menu[] = []
 
+  filterRecipeName: string = ""
+
   displayedColumns: string[] = ['recipe_name', 'ingredients', 'description', 'remain_order', 'price', 'status', 'actions'];
   dataSource = new MatTableDataSource([])
 
@@ -26,13 +28,34 @@ export class MenuManagementComponent implements OnInit {
     private dialog: MatDialog
   ) { }
 
+  // Accent(value: string): string {
+  //   let accent:string[] = value.split('');
+  //   accent = accent.map(data => {
+  //     const splited = data.normalize('NFD').split('');
+  //     return splited[0]
+  //   })
+  //   return accent.join('');
+  // }
+
+  // filterByRecipeName() {
+  //   let fil = this.menu
+  //   fil = fil.filter(data => {
+  //     let name = data.recipe_name;
+  //     name = name.toLowerCase();
+  //     name = name.replace(/[.\s,]/g, '');
+  //     name = this.Accent(name)
+  //     return name.includes(this.Accent(this.filterName.toLowerCase().replace(/[.\s,]/g, '')))
+  //   })
+  //   this.dataSource.data = fil
+  // }
+
   ngOnInit(): void {
     this.init(this.paginator)
   }
 
-  init(paginationObj:any) {
+  init(paginationObj: any) {
 
-    const pagination : any = {
+    const pagination: any = {
       page: paginationObj?.page ?? 1,
       limit: paginationObj?.limit ?? 5,
     }
@@ -40,7 +63,7 @@ export class MenuManagementComponent implements OnInit {
     this.subs.sink = this.menuService.getRecipe(pagination).subscribe(resp => {
 
       this.paginator.length = resp.data.GetAllRecipes.count;
-      
+
       // this.paginator.pageSize = this.pageSizeOptions[0];
 
       this.menu.push(resp.data.GetAllRecipes.data_recipes)
@@ -63,29 +86,33 @@ export class MenuManagementComponent implements OnInit {
           title: 'Success',
           text: 'Your work has been saved',
         });
-        this.init(this.paginator)
+        this.init(true)
       })
       console.log('The dialog was closed')
     }
     );
   }
 
+  a:any
 
   onEdit(parameter: any) {
     console.log(parameter);
-    const bebas : any = {
-      ...parameter,ingredients : parameter.ingredients.map((ingredient: any) =>{
-        return {ingredient_id : ingredient.ids.id, stock_used : ingredient.stock_used, name: ingredient.ids.name}
-    })
+    const bebas: any = {
+      ...parameter, ingredients: parameter.ingredients.map((ingredient: any) => {
+        return { ingredient_id: ingredient.ids.id, stock_used: ingredient.stock_used, name: ingredient.ids.name }
+      })
     }
     console.log(bebas);
-    
+
     const dialogRef = this.dialog.open(MenuManagementUpdateComponent, {
       width: '100%', height: '100%',
       data: bebas || null
     });
 
     dialogRef.afterClosed().subscribe(result => {
+      if(result){
+        this.init(result)
+      }
       console.log('The dialog was closed');
     }
 
@@ -93,22 +120,75 @@ export class MenuManagementComponent implements OnInit {
   }
 
   onDelete(parameter: any) {
-    this.menuService.deleteRecipe(parameter).subscribe(resp => {
-      this.init(this.paginator)
+    Swal.fire({
+      title: 'Are you sure?',
+      text: "this menu will be permanently deleted!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Yes, delete now!'
+    }).then((result:any) => {
+      if (result.isConfirmed){
+        this.menuService.deleteRecipe(parameter).subscribe(resp => {
+          Swal.fire({
+            title: 'Deleted!',
+            text: "this menu has been deleted",
+            icon: 'success',
+        })
+          this.init(true)
+        })
+      }
     })
+    
     console.log(parameter);
   }
+
+
+
+  checked = false;
+
+  
+  onChanged(event: any, element: any) {
+    const check = {
+      id: element.id,
+      status: event.checked ? 'active' : 'unpublish'
+    };
+    // if(check.status === 'active'){
+    //   check.status = 'active'
+    // }else if(check.status === 'unpublish'){
+    //   check.status = 'unpublish'
+    // }
+    // Swal.fire({
+    //   title: 'Are you sure want change status to ' + check.status + '?',
+    //   showDenyButton: false,
+    //   showCancelButton: true,
+    //   showConfirmButton: true,
+    //   denyButtonText: `Yes`,
+    // }).then((result)=>{
+    //   if(result.isConfirmed){
+    //     this.menuService.updatepublish(check).subscribe(() => {
+    //       Swal.fire({
+    //         title: 'you have been change status to ' + check.status
+    //       })
+    //       this.init(true)
+    //     })
+    //   }
+    // })                                  
+  }
+
+
 
   @ViewChild('paginator') paginator!: MatPaginator
 
   pageSizeOptions: number[] = [5];
 
-  onPagination(event:any) {
+  onPagination(event: any) {
     console.log(event);
-    
-    const pagination ={
-      limit : event?.pageSize,
-      page :  event?.pageIndex+1
+
+    const pagination = {
+      limit: event?.pageSize,
+      page: event?.pageIndex + 1
     }
     this.init(pagination)
   }
